requirements.txt
fastapi==0.115.0
numpy==1.22.2
python-multipart==0.0.10
Pillow==10.4.0
torch==1.13.1
torch_neuron==1.13.1.2.11.7.0
torchvision==0.14.1
uvicorn==0.30.6
